Vendor prefix shim
Enabled capturing of event types that are prefixed. For example, if you
are running in Chrome and the 'click' event is attached, the
'webkitclick' event will also be attached.

While this behavior isn't particularly useful for 'click' events, it can
extend the capturing of event types that are incubating. At the time of
authoring the 'visibilychange' event is supported in every major browser
with a prefix (Opera supports it without a prefix). This shim allows the
event to be defined according to the spec, while draft implementations
can be captured without needing to respecify the event definition.

We do not attempt to white/black-list prefixed events. Maintaining that
list would be a nightmare. Rather this shim treats every event definition
as potentially prefixed. The overhead of listening for events that will
never fire should be negligible.

Issue: #6